+++ mycodo/controllers/controller_output.py
@@ -397,18 +397,23 @@ class OutputController(AbstractController, threading.Thread):
         except Exception:
             self.logger.exception("is_setup() exception")
 
-    def call_module_function(self, button_id, args_dict, unique_id=None, thread=True):
+    def call_module_function(self, button_id, args_dict, unique_id=None, thread=True, return_from_function=False):
         """Execute function from custom action button press."""
         try:
             run_command = getattr(self.output[unique_id], button_id)
-            if thread:
+            if not thread or return_from_function:
+                return_val = run_command(args_dict)
+                if return_from_function:
+                    return 0, return_val
+                else:
+                    return 0, f"Command sent to Output Controller. Returned: {return_val}"
+            else:
                 thread_run_command = threading.Thread(
                     target=run_command,
                     args=(args_dict,))
                 thread_run_command.start()
                 return 0, "Command sent to Output Controller and is running in the background."
-            else:
-                return_val = run_command(args_dict)
-                return 0, f"Command sent to Output Controller. Returned: {return_val}"
-        except:
-            self.logger.exception(f"Error executing custom action '{button_id}'")
+        except Exception as err:
+            msg = f"Error executing function '{button_id}': {err}"
+            self.logger.exception(msg)
+            return 1, msg