+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - Add I2C interface for K30 Input
  - Add Day to unit conversion
  - Add ability to disable any Input/Output/Function/Widget option (user can view but not change)
+ - Add ability to return Input/Output/Function object with module_function()
 
 
 ## 8.15.8 (2023-04-06)